requirements.txt
dv
tqdm
numba
numpy
scipy
pyyaml
pandas
tabulate
opencv-python